requirements.txt
requests
pandas
python-dateutil
pytest
textual>=8.0
tomli-w
openpyxl
pytest-asyncio
